+++ crates/cadmus/src/main.rs
@@ -1,7 +1,7 @@
 mod app;
 
 use crate::app::run;
-use plato_core::anyhow::Error;
+use cadmus_core::anyhow::Error;
 
 fn main() -> Result<(), Error> {
     run()?;