+++ crates/emulator/src/main.rs
@@ -1,40 +1,42 @@
-use plato_core::anyhow::{Context as ResultExt, Error};
-use plato_core::battery::{Battery, FakeBattery};
-use plato_core::chrono::Local;
-use plato_core::color::Color;
-use plato_core::context::Context;
-use plato_core::device::CURRENT_DEVICE;
-use plato_core::document::sys_info_as_html;
-use plato_core::font::Fonts;
-use plato_core::framebuffer::{Framebuffer, UpdateMode};
-use plato_core::frontlight::{Frontlight, LightLevels};
-use plato_core::geom::{Axis, Rectangle};
-use plato_core::gesture::{gesture_events, GestureEvent};
-use plato_core::helpers::{load_toml, save_toml};
-use plato_core::input::{ButtonCode, ButtonStatus, DeviceEvent, FingerStatus};
-use plato_core::library::Library;
-use plato_core::lightsensor::LightSensor;
-use plato_core::png;
-use plato_core::pt;
-use plato_core::settings::{IntermKind, Settings, SETTINGS_PATH};
-use plato_core::view::calculator::Calculator;
-use plato_core::view::common::{
+use cadmus_core::anyhow::{Context as ResultExt, Error};
+use cadmus_core::battery::{Battery, FakeBattery};
+use cadmus_core::chrono::Local;
+use cadmus_core::color::Color;
+use cadmus_core::context::Context;
+use cadmus_core::device::CURRENT_DEVICE;
+use cadmus_core::document::sys_info_as_html;
+use cadmus_core::font::Fonts;
+use cadmus_core::framebuffer::{Framebuffer, UpdateMode};
+use cadmus_core::frontlight::{Frontlight, LightLevels};
+use cadmus_core::geom::{Axis, Rectangle};
+use cadmus_core::gesture::{gesture_events, GestureEvent};
+use cadmus_core::helpers::{load_toml, save_toml};
+use cadmus_core::input::{ButtonCode, ButtonStatus, DeviceEvent, FingerStatus};
+use cadmus_core::library::Library;
+use cadmus_core::lightsensor::LightSensor;
+use cadmus_core::png;
+use cadmus_core::pt;
+use cadmus_core::settings::{IntermKind, Settings, SETTINGS_PATH};
+use cadmus_core::view::calculator::Calculator;
+use cadmus_core::view::common::{
     locate, locate_by_id, overlapping_rectangle, transfer_notifications,
 };
-use plato_core::view::common::{toggle_input_history_menu, toggle_keyboard_layout_menu};
-use plato_core::view::dialog::Dialog;
-use plato_core::view::dictionary::Dictionary;
-use plato_core::view::frontlight::FrontlightWindow;
-use plato_core::view::home::Home;
-use plato_core::view::intermission::Intermission;
-use plato_core::view::menu::{Menu, MenuKind};
-use plato_core::view::notification::Notification;
-use plato_core::view::reader::Reader;
-use plato_core::view::rotation_values::RotationValues;
-use plato_core::view::sketch::Sketch;
-use plato_core::view::touch_events::TouchEvents;
-use plato_core::view::{handle_event, process_render_queue, wait_for_all, RenderData, RenderQueue};
-use plato_core::view::{AppCmd, EntryId, EntryKind, Event, View, ViewId};
+use cadmus_core::view::common::{toggle_input_history_menu, toggle_keyboard_layout_menu};
+use cadmus_core::view::dialog::Dialog;
+use cadmus_core::view::dictionary::Dictionary;
+use cadmus_core::view::frontlight::FrontlightWindow;
+use cadmus_core::view::home::Home;
+use cadmus_core::view::intermission::Intermission;
+use cadmus_core::view::menu::{Menu, MenuKind};
+use cadmus_core::view::notification::Notification;
+use cadmus_core::view::reader::Reader;
+use cadmus_core::view::rotation_values::RotationValues;
+use cadmus_core::view::sketch::Sketch;
+use cadmus_core::view::touch_events::TouchEvents;
+use cadmus_core::view::{
+    handle_event, process_render_queue, wait_for_all, RenderData, RenderQueue,
+};
+use cadmus_core::view::{AppCmd, EntryId, EntryKind, Event, View, ViewId};
 use sdl2::event::Event as SdlEvent;
 use sdl2::keyboard::{Keycode, Mod, Scancode};
 use sdl2::mouse::MouseState;
@@ -50,7 +52,7 @@ use std::sync::mpsc;
 use std::thread;
 use std::time::Duration;
 
-pub const APP_NAME: &str = "Plato";
+pub const APP_NAME: &str = "Cadmus";
 const DEFAULT_ROTATION: i8 = 1;
 
 const CLOCK_REFRESH_INTERVAL: Duration = Duration::from_secs(60);
@@ -260,7 +262,7 @@ fn main() -> Result<(), Error> {
     let video_subsystem = sdl_context.video().unwrap();
     let (width, height) = CURRENT_DEVICE.dims;
     let window = video_subsystem
-        .window("Plato Emulator", width, height)
+        .window("Cadmus Emulator", width, height)
         .position_centered()
         .build()
         .unwrap();
@@ -661,7 +663,7 @@ fn main() -> Result<(), Error> {
                     let dialog = Dialog::new(
                         ViewId::AboutDialog,
                         None,
-                        format!("Plato {}", env!("GIT_VERSION")),
+                        format!("Cadmus {}", env!("GIT_VERSION")),
                         &mut context,
                     );
                     rq.add(RenderData::new(