+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Changelog
 
-## [0.9.45](https://github.com/OGKevin/plato/compare/v0.9.44...v0.9.45) (2026-01-12)
+## Cadmus Fork
+
+This project is now maintained as **Cadmus**, a fork of the [Plato](https://github.com/baskerville/plato) document reader.
+
+## [0.9.45](https://github.com/OGKevin/cadmus/compare/v0.9.44...v0.9.45) (2026-01-12)
 
 
 ### Features
 
-* add restart application menu option ([#8](https://github.com/OGKevin/plato/issues/8)) ([4cf8af1](https://github.com/OGKevin/plato/commit/4cf8af12a03ecd7c74e86c575c7c84dfe51df358))
+* add restart application menu option ([#8](https://github.com/OGKevin/cadmus/issues/8)) ([4cf8af1](https://github.com/OGKevin/cadmus/commit/4cf8af12a03ecd7c74e86c575c7c84dfe51df358))
 
 
 ### Bug Fixes
 
-* **fetcher:** add https support ([#39](https://github.com/OGKevin/plato/issues/39)) ([58b64f9](https://github.com/OGKevin/plato/commit/58b64f9a666cf52300a70a4331960b6e4e94abcc))
+* **fetcher:** add https support ([#39](https://github.com/OGKevin/cadmus/issues/39)) ([58b64f9](https://github.com/OGKevin/cadmus/commit/58b64f9a666cf52300a70a4331960b6e4e94abcc))